jira-client/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jira-client"
version = "0.1.0"
description = "A Python library for Atlassian Jira Cloud operations"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["jira", "atlassian", "project-management", "api-client"]
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
